fix camera block in published proj
